Place edit button within account menu buttons
