+++ packages/page-accounts/src/Sidebar/AccountMenuButtons.tsx
@@ -1,0 +1,136 @@
+// Copyright 2017-2021 @polkadot/app-accounts authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+import React, { useCallback } from 'react';
+import styled from 'styled-components';
+
+import Transfer from '@polkadot/app-accounts/modals/Transfer';
+import { useTranslation } from '@polkadot/app-accounts/translate';
+import { Button } from '@polkadot/react-components';
+import { useToggle } from '@polkadot/react-hooks';
+import { AddressFlags } from '@polkadot/react-hooks/types';
+
+interface Props {
+  className?: string;
+  flags: AddressFlags;
+  isEditingName: boolean;
+  isEditing: boolean;
+  toggleIsEditingName: () => void;
+  toggleIsEditingTags: () => void;
+  onCancel: () => void;
+  onSaveName: () => void;
+  onSaveTags: () => void;
+  onForgetAddress: () => void;
+  onUpdateName: () => void;
+  recipientId: string;
+}
+
+function AccountMenuButtons ({ className = '', flags, isEditing, isEditingName, onCancel, onForgetAddress, onSaveName, onSaveTags, onUpdateName, recipientId, toggleIsEditingName, toggleIsEditingTags }: Props) {
+  const { t } = useTranslation();
+
+  const _onForgetAddress = useCallback(
+    (): void => {
+      onForgetAddress();
+      onUpdateName && onUpdateName();
+    },
+    [onForgetAddress, onUpdateName]
+  );
+
+  const [isTransferOpen, toggleIsTransferOpen] = useToggle();
+
+  const updateTags = useCallback(() => {
+    onSaveTags();
+  }, [onSaveTags]);
+
+  const toggleIsEditing = useCallback(() => {
+    flags.isEditable && toggleIsEditingName();
+    toggleIsEditingTags();
+  }, [flags.isEditable, toggleIsEditingName, toggleIsEditingTags]);
+
+  const _onUpdateName = useCallback(
+    (): void => {
+      onSaveName();
+      onUpdateName && onUpdateName();
+    },
+    [onSaveName, onUpdateName]
+  );
+
+  const updateName = useCallback(() => {
+    if (isEditingName && (flags.isInContacts || flags.isOwned)) {
+      _onUpdateName();
+      toggleIsEditingName();
+    }
+  }, [isEditingName, flags.isInContacts, flags.isOwned, _onUpdateName, toggleIsEditingName]);
+
+  const onEdit = useCallback(() => {
+    if (isEditing) {
+      updateName();
+      updateTags();
+    }
+
+    toggleIsEditing();
+  }, [isEditing, toggleIsEditing, updateName, updateTags]);
+
+  return (
+    <div className={`${className} ui--AddressMenu-buttons ${isEditing ? 'isEditing' : ''}`}>
+      {isEditing
+        ? (
+          <Button.Group>
+            <Button
+              icon='times'
+              label={t<string>('Cancel')}
+              onClick={onCancel}
+            />
+            <Button
+              icon='save'
+              label={t<string>('Save')}
+              onClick={onEdit}
+            />
+          </Button.Group>
+        )
+        : (
+          <Button.Group>
+            <Button
+              icon='paper-plane'
+              isDisabled={isEditing}
+              label={t<string>('Send')}
+              onClick={toggleIsTransferOpen}
+            />
+            {!flags.isOwned && !flags.isInContacts && (
+              <Button
+                icon='plus'
+                isDisabled={isEditing}
+                label={t<string>('Save')}
+                onClick={_onUpdateName}
+              />
+            )}
+            {!flags.isOwned && flags.isInContacts && (
+              <Button
+                icon='ban'
+                isDisabled={isEditing}
+                label={t<string>('Remove')}
+                onClick={_onForgetAddress}
+              />
+            )}
+            <Button
+              icon='edit'
+              label={t<string>('Edit')}
+              onClick={onEdit}
+            />
+          </Button.Group>)}
+      {isTransferOpen && (
+        <Transfer
+          key='modal-transfer'
+          onClose={toggleIsTransferOpen}
+          recipientId={recipientId}
+        />
+      )}
+    </div>
+  );
+}
+
+export default React.memo(styled(AccountMenuButtons)`
+  .ui--Button-Group {
+    margin-bottom: 0;
+  }
+`);
